fix stack overflow when traversing large codebases
Running "npx phero server build" on a sufficiently large project will
result in the following error:

    RangeError: Maximum call stack size exceeded

This is due to the "recursivelyFindThrowStatements" function, which
creates a new stack frame for every iteration as it traverses the
codebase. This is despite it using tail recursion, because v8 simply
does not implement tail-call optimization.

Fix this by rewriting this function to not use recursion at all.