fix errors from void response RPCs failing to parse
For RPCs that return void, we try to parse to error object from the
response data directly, instead of the "error" property where they
actually live. This was fixed before for non-void RPCs in commit
1a70b8a (fix errors from exported servers failing to parse, 2023-08-04),
but never got fixed for void RPCs.

This solves that by unifying the error handling of the void and non-void
cases, making it harder for the logic to get out-of-sync.